Switching to a non-commercial license now (see extended desc for why)
Basically I want to address 3 points about the MIT license:
- Freely publishing:
Still able to publish this e.g. in a (YouTube) video or blog post, but you just have to attribute it accordingly.
- Redistributing:
Allowed, but only the original work can be publicly redistributed. This means you can't make a fork, make sufficiently significant changes (to the point where output results/data differ from the original code, unless the clause under claim of a bug fix or error in the code causes a defect to what true values should actually be for a given tool/calculation) and then publicly publish that fork. This is to protect the objective truth of what these tools are calculating, as if a modified copy introduces a bug or modification to the logic that now misaligns it from the actual reality of the game, then users with possession of this inaccurate fork may use the program unbeknownst to this and then they'd be working with incorrect data.
- Commercial resale:
This is just an obvious no no no, like if I'm not making any money of this hobby project, then anyone else's sorry a** ain't either